requirements.txt
anyascii==0.3.3
asgiref==3.11.1
asttokens==3.0.1
beautifulsoup4==4.12.3
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
decorator==5.2.1
defusedxml==0.7.1
Django==5.1.15
django-autoslug==1.9.9
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
django-extensions==4.1
django-filter==24.3
django-modelcluster==6.4.1
django-permissionedforms==0.1
django-taggit==6.1.0
django-treebeard==4.8.0
djangorestframework==3.16.1
draftjs_exporter==5.2.0
et_xmlfile==2.0.0
executing==2.2.1
filetype==1.2.0
idna==3.11
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
l18n==2021.3
laces==0.1.2
matplotlib-inline==0.2.1
mysqlclient==2.2.8
openpyxl==3.1.5
parso==0.8.6
pillow==11.3.0
pillow_heif==1.2.0
prompt_toolkit==3.0.52
pure_eval==0.2.3
Pygments==2.19.2
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
stack-data==0.6.3
telepath==0.3.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wagtail==6.3
wcwidth==0.6.0
Willow==1.12.0
django-allauth==65.14.3
pudb==2025.1.5